+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to
 - @cosmjs/stargate: Add missing exports `AminoMsgTransfer`/`isAminoMsgTransfer`.
 - @cosmjs/stargate: Add support for `MsgVoteWeighted` (register by default and
   create Amino JSON converters) ([#1224]).
+- @cosmjs/stargate: Add Amino JSON support for `MsgCreateVestingAccount`.
 
 [#1072]: https://github.com/cosmos/cosmjs/issues/1072
 [#1154]: https://github.com/cosmos/cosmjs/issues/1154